Stub of cells-based rendering.
Text data is rendered into an array of cells which allow to abstract
away fine clipping, horizontal scroll and most importantly allow to
simplify UTF-8 handling.

The design includes a linear buffer (arena pool) which is used for large
Unicode sequences (> 8 bytes).

This commit introduces a few regressions (fine clipping and horizontal
scroll) but also fixes a few bugs. It's still a work in progress but it
should be a very solid and future-proof approach to rendering.