Add fine clipping, horizontal scroll and some fix.
Better decoupling between the core and the backend. We should be able to
think in terms of "unit points" for the cells instead of TUI columns.
For example in an SDL backend we can think of "width" as pixels without
having to change anything into the core. This is the theory.